LPS-69953 Move to StagingGroupHelper it belongs there
