+++ modules/apps/staging/staging-api/src/main/java/com/liferay/staging/StagingGroupHelper.java
@@ -72,6 +72,9 @@ public interface StagingGroupHelper {
 
 	public boolean isStagedPortlet(long groupId, String portletId);
 
+	public boolean isStagedPortletData(long groupId, String className)
+		throws Exception;
+
 	public boolean isStagingGroup(Group group);
 
 	public boolean isStagingGroup(long groupId);
