+++ modules/apps/staging/staging-impl/build.gradle
@@ -8,5 +8,6 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly project(":apps:changeset:changeset-api")
+	compileOnly project(":apps:export-import:export-import-api")
 	compileOnly project(":apps:staging:staging-api")
 }
